Extend inheritance reordering to interface / type generation. Extend tests to model generation.
