+++ packages/odata2ts/test/generator/v2/EntityBasedGenerationTests.ts
@@ -164,6 +164,24 @@ export function createEntityBasedGenerationTests(
     });
   });
 
+  test(`${testSuiteName}: base class order`, async () => {
+    // given an entity hierarchy not in order of inheritance
+    odataBuilder
+      .addEntityType("Child", "Parent", (builder) =>
+        builder.addKeyProp("id2", ODataTypesV2.Boolean).addProp("Ch1ld1shF4n", ODataTypesV2.Boolean)
+      )
+      .addEntityType("GrandParent", undefined, (builder) => builder.addKeyProp("id", ODataTypesV2.Boolean))
+      .addEntityType("Parent", "GrandParent", (builder) => builder.addProp("parentalAdvice", ODataTypesV2.Boolean));
+
+    // when generating model
+    // then match fixture text from hierachy in order
+    await generateAndCompare("baseClassOrder", "entity-hierarchy.ts", {
+      skipIdModels: false,
+      skipEditableModels: false,
+      disableAutoManagedKey: true,
+    });
+  });
+
   test(`${testSuiteName}: entity & enum`, async () => {
     // given an entity with enum props
     odataBuilder
